Conflicts:

	www/public_html/administrator/project_types/templates.html
	www/public_html/employee/production/print_overview.html
	www/public_html/includes/main/summ/paper_dropdowns.html

+++ perl/openprint/Pricelist.pm
@@ -12,8 +12,12 @@ require openprint::logs;
 use openprint ();
 
 my $debug = 1;
-use vars qw( %fields );
+use vars qw( $log $dbh $table $serial %fields %transforms %defaults );
+*log = \$openprint::log;
+*dbh = \$openprint::dbh;
 
+$table = 'pricelists';
+$serial = 'pricelists_id_seq';
 %fields = (
 	'id'			=>	'id',
 	'name'			=>	'name',
@@ -47,46 +51,20 @@ sub find {
 
 } # end sub find
 
-sub load {
-	my ( $self, $data ) = @_;
-	if ( ! $data ) {
-		$data = $openprint::dbh->selectrow_hashref( 'SELECT * FROM Pricelists WHERE id=?', {}, $$self{'id'} );
-	} # end if
-	@$self{keys %fields} = @$data{@fields{keys %fields}};
-} # end sub load
-
 sub delete {
 	my $self = shift;
 
+	my @PaperPrices = $self->getPrices('Paper');
 	my $ac = sql::start_transaction( $openprint::dbh );
+	sql::update( undef, undef, 'companies', ['pricelist_id=?', $$self{'id'}], 'pricelist_id', undef );
     sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM Service_Prices WHERE pricelist_id=?}, $$self{id} );
     sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM tbl_Material_Prices WHERE lngListIndex=?}, $$self{id} );
-    sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM Paper_Prices WHERE lngListIndex=?}, $$self{id} );
+    sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM Paper_Prices WHERE lngListIndex=?}, $$self{id} ) if @PaperPrices;
     sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM Product_Prices WHERE pricelist_id=?}, $$self{id} );
     sql::execute( $openprint::log, $openprint::dbh, q{DELETE FROM Pricelists WHERE id=?}, $$self{id} );
 	sql::end_transaction( $openprint::dbh, $ac );
 	openprint::logs::insertLogRecord('9', "Price List Index: " . $$self{id},);
 } # end sub delete
-
-sub save {
-	my ( $self, $param ) = @_;
-	if ( $param ) {
-		$$self{'name'} = $$param{'Name'};
-		$$self{'description'} = $$param{'Description'};
-		$$self{'currency_id'} = $$param{'Currency'};
-	} # end if
-	my @sql = (
-		'name', $$self{'name'}, 'Description', $$self{'description'}, 'currency_id', $$self{'currency_id'},
-);
-	if ( ! $$self{'id'} ) {
-		@$self{'id'} = sql::execute( $openprint::log, $openprint::dbh, q{SELECT nextval('pricelists_id_seq'::text)} );
-		sql::insert( $openprint::log, $openprint::dbh, 'Pricelists', 'id', $$self{'id'}, @sql );
-		openprint::logs::insertLogRecord('30', "Price List Index: " . $$self{'id'} . " - " . $$self{'name'},);
-	} else { #save
-		sql::update( $openprint::log, $openprint::dbh, 'Pricelists', ['Id=?', $$self{'id'}],@sql );
-		openprint::logs::insertLogRecord('31', "Price List Index: " . $$self{'id'} . " - " . $$self{'name'},);
-	} # end if
-} # end sub save
 
 sub getPrices {
 	my ( $self, $type ) = @_;
